# List Functor: mix unrolled and reverse map