--- a/bower.json
+++ b/bower.json
@@ -31,6 +31,7 @@
     "purescript-arrays": "^4.0.0",
     "purescript-assert": "^3.0.0",
     "purescript-console": "^3.0.0",
-    "purescript-math": "^2.0.0"
+    "purescript-math": "^2.0.0",
+    "purescript-minibench": "matthewleon/purescript-minibench#gc"
   }
 }