--- a/src/Data/List/Types.purs
+++ b/src/Data/List/Types.purs
@@ -1,4 +1,8 @@
-module Data.List.Types where
+module Data.List.Types
+  ( List(..)
+  , (:)
+  , NonEmptyList(..)
+  ) where
 
 import Prelude
 
@@ -67,7 +71,47 @@ instance monoidList :: Monoid (List a) where
   mempty = Nil
 
 instance functorList :: Functor List where
-  map f = foldr (\x acc -> f x : acc) Nil
+  map = listMap
+
+-- chunked list Functor inspired by OCaml
+-- https://discuss.ocaml.org/t/a-new-list-map-that-is-both-stack-safe-and-fast/865
+-- chunk sizes determined through experimentation
+listMap :: forall a b. (a -> b) -> List a -> List b
+listMap f = startUnrolledMap unrollLimit
+  where
+  -- iterate the unrolled map up to 1000 times,
+  -- which hits up to 5000 elements
+  unrollLimit = 1000
+
+  startUnrolledMap :: Int -> List a -> List b
+  startUnrolledMap 0 (x : xs) = f x : chunkedRevMap xs
+  startUnrolledMap n (x1 : x2 : x3 : x4 : x5 : xs) =
+    f x1 : f x2 : f x3 : f x4 : f x5 : startUnrolledMap (n - 1) xs
+  startUnrolledMap n (x1 : x2 : x3 : x4 : xs) =
+    f x1 : f x2 : f x3 : f x4 : startUnrolledMap (n - 1) xs
+  startUnrolledMap n (x1 : x2 : x3 : xs) =
+    f x1 : f x2 : f x3 : startUnrolledMap (n - 1) xs
+  startUnrolledMap n (x1 : x2 : xs) =
+    f x1 : f x2 : startUnrolledMap (n - 1) xs
+  startUnrolledMap n (x : xs) = f x : startUnrolledMap (n - 1) xs
+
+  startUnrolledMap _ Nil = Nil
+
+  chunkedRevMap :: List a -> List b
+  chunkedRevMap = go Nil
+    where
+    go :: List (List a) -> List a -> List b
+    go chunksAcc chunk@(x1 : x2 : x3 : x4 : x5 : xs) =
+      go (chunk : chunksAcc) xs
+    go chunksAcc finalChunk =
+      reverseUnrolledMap chunksAcc $ startUnrolledMap 0 finalChunk
+
+    reverseUnrolledMap :: List (List a) -> List b -> List b
+    reverseUnrolledMap ((x1 : x2 : x3 : _) : cs) acc =
+      reverseUnrolledMap cs (f x1 : f x2 : f x3 : acc)
+    -- if we pattern match on Nil, we need a Partial constraint,
+    -- which kills TCO
+    reverseUnrolledMap _ acc = acc
 
 instance functorWithIndexList :: FunctorWithIndex Int List where
   mapWithIndex f = foldrWithIndex (\i x acc -> f i x : acc) Nil