--- a/test/Test/Data/List.purs
+++ b/test/Test/Data/List.purs
@@ -363,6 +363,9 @@ testList = do
   log "map should be stack-safe"
   void $ pure $ map id (1..100000)
 
+  log "map should be correct"
+  assert $ (1..1000000) == map (_ + 1) (0..999999)
+
   log "transpose"
   assert $ transpose (l [l [1,2,3], l[4,5,6], l [7,8,9]]) ==
                      (l [l [1,4,7], l[2,5,8], l [3,6,9]])