@matthewleon 2017-11-21
Will update this with details on my machine setup.