@matthewleon 2017-11-22
> straight traversal of the spine using a mutable cons cell, a la Scheme

I will research this and see if I can come up with something crafty. Is the idea here that we jump into `ST`?