@paf31 2017-11-22
I think it would be FFI instead of ST unfortunately, unless we can come up with a way to modify cons cells in `ST`, which would mean some sort of weird `STList` data type.