@matthewleon 2017-11-22
> unless we can come up with a way to modify cons cells in ST, which would mean some sort of weird STList data type.

I was thinking along those lines. I'll try it and benchmark it.