@matthewleon 2017-11-22
Though upon further thought, I don't see how to do that without running into the problem of stack-safe recursion on `Eff` being very slow. Perhaps the saner thing to do is to try the mutable cons cell in FFI... I'll have to use JS anyway to optimize the unroll.