@matthewleon 2017-11-22
Further, outside of `ST`, I'm having trouble figuring out how to write JavaScript for this that isn't dependent on the way the compiler chooses to structure the `List` type, which seems very brittle.