@paf31 2017-11-23
Yes, agreed, it doesn't seem worth it to make the code non-portable.