@matthewleon 2017-11-23
> Yes, agreed, it doesn't seem worth it to make the code non-portable.

So do you prefer that the code be largely left as-is, because this version is portable and has much better performance than what we had before, at least in Node?

Or should I go ahead and try to do the mutable cons cell + unrolling in FFI JavaScript?

The answer to me really isn't obvious, as I'm unaware of what PS's stance regarding library portability is. As it stands, this lib doesn't use any FFI at all – directly. I believe that `Lazy`, however, uses FFI. Which makes at least the lazy lists non-portable, no?