@paf31 2017-11-23
Well the lazy data structure doesn't rely on the compiler internals in any way.

Here is a quick FFI experiment:

```javascript
exports.mapList = function(f) {
  return function(l) {
    if (l.constructor.name === 'Nil') {
      return l;
    } else {
      var nil = require('Data.List.Types').Nil.value;
      var cons = require('Data.List.Types').Cons;
      var result = new cons(null, null);

      var input = l;
      var output = result;

      while (true) {
        output.value0 = f(input.value0);
        output.value1 = new cons(null, null);
        
        if (input.value1.constructor.name == 'Nil') {
          break;
        }

        input = input.value1;
        output = output.value1;
      }

      output.value1 = nil;

      return result;
    }
  }
}
```

It gets about 3ms on a list of 100000 elements on my machine, but I'd be interested to hear what you get.

So it seems like the loop unrolling is really helping quite a bit here.