@paf31 2017-11-23
I've put my code in a repo [here](https://github.com/paf31/purescript-lists-fast), along with `zipWith` and `filter` replacements.