@matthewleon 2017-11-23
Cool. To be sure, those expected speedups might decrease a fair bit once we
merge in non-FFI optimizations to Data.List.

On Thu, Nov 23, 2017 at 4:57 PM, Phil Freeman <notifications@github.com>
wrote:

> I've put my code in a repo here
> <https://github.com/paf31/purescript-lists-fast>, along with zipWith and
> filter replacements.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/purescript/purescript-lists/pull/135#issuecomment-346702273>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AARBS-HDNYnlzTOJnDqjiDyI1e28Lo4zks5s5epYgaJpZM4QmhW9>
> .
>