@matthewleon 2017-11-25
I've added benchmarking to this branch.