@paf31 2017-11-30
Thanks!

What do other people think? @garyb @hdgarrood 

Especially about the potential stack overflow issue mentioned above.