@garyb 2017-11-30
@paf31 what was the stack overflow issue? Just that thing that if the stack is near-overflow already it might push it over the edge?