@hdgarrood 2017-11-30
Yeah, if `f` makes enough recursive calls on average, I think that could cause stack overflows. This looks really good though! If you could address @paf31's recent comments and do a little testing to check that stack-safety isn't significantly worse than it is now, I'm :+1: to merge.

Come to think of it, do we have any testing for stack overflows in `map` in this library already?