@paf31 2017-11-30
To be clear, the issue is that we go from "`map` never grows the stack" to "`map` always grows the stack, but maybe not by very much". The growth should be linear in the size of the list.