@natefaubion 2017-11-30
With this, isn't it "map grows the stack by a fixed maximum amount"?