@paf31 2017-11-30
Well `startUnrolledMap` calls itself recursively, but not tail-recursively, so no, I think the stack usage grows with the size of the input.