@natefaubion 2017-11-30
Right, but it has a maximum threshold defined be `unrollLimit`, which causes it to fallback to the slower, stack-safe variant which halts the growth. So at most, `map` itself will consume 1000 frames .