@safareli 2017-11-30
If frame limit is 10,000 and before map is called 9,001 frames are used then using 1k frames will overflow the stack right? is this fine?