@paf31 2017-11-30
Nested `map`s are also quite common. I don't think this is enough to cause a problem on its own, but it could worsen it. The question for me is not whether this code should be included, but whether it should be the default. Do we want a default which works in every case but is slow in most cases, or one which works in most cases but overflows the stack in a tiny number of cases?

I think it's probably okay to pick this as the default, but I just want to make sure the trade off is clear now as we decide.