@matthewleon 2017-11-30
> There's also the question of what the stack limit is. It does 5 items at a time, so it will eat stack up to 5000 items. Maybe we don't need to handle that many, and we can reduce it to 200 frames if it's a concern?

I wouldn't mind doing that. Maybe it's worth me doing some exploratory tests to see how much breathing room Node gives us?