@matthewleon 2017-11-30
regarding the call stack size: http://2ality.com/2014/04/call-stack-size.html