@matthewleon 2017-11-30
> we can reduce it to 200 frames if it's a concern?

I've made a commit that does this. I figure that this means we still get a big speed-up for common List use-cases, while if people are serious about having faster iteration through longer lists, they can use the fast list repo.