@paf31 2017-12-10
Given the possible stack-safety issue, what do people think about providing the naive (safe) implementation as a separate named function? Or vice versa?