@matthewleon 2017-12-11
One possibility is to just use the stack-safe unrolled operation, which should still give a speed boost, and recommend that people turn to `purescript-lists-fast` if speed is important.