@paf31 2017-12-11
Would it be possible to try that and compare the performance easily?