@matthewleon 2017-12-12
I made a branch that only uses the stack-safe reverse map and... It's actually faster.

map: list (2000 elems)
mean   = 53.33 μs -> 31.24 μs

map: list (10000 elems)
mean   = 107.73 μs -> 98.17 μs

I've merged it in.