@matthewleon 2017-12-12
Just realized I can also inline the `unrolledMap`, will do that and push.