@matthewleon 2017-12-12
Actually, inlining `unrolledMap` slows things down significantly. Mysteries of JS... This should be good for review.